src/backend/requirements.txt
fastapi==0.116.1
psycopg2-binary==2.9.10
pydantic==2.11.7
pyjwt[crypto]==2.3.0
SQLAlchemy==2.0.43
uvicorn==0.35.0
